requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
feedparser>=6.0.11
pytrends>=4.9.2
google-genai>=1.11.0
pytest>=8.3.0
